Migrating `InputStateBadge` to TS.
