--- graylog2-web-interface/src/components/inputs/InputStateBadge.jsx
@@ -1,133 +1,0 @@
-/*
- * Copyright (C) 2020 Graylog, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the Server Side Public License, version 1,
- * as published by MongoDB, Inc.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * Server Side Public License for more details.
- *
- * You should have received a copy of the Server Side Public License
- * along with this program. If not, see
- * <http://www.mongodb.com/licensing/server-side-public-license>.
- */
-import PropTypes from 'prop-types';
-import React from 'react';
-// eslint-disable-next-line no-restricted-imports
-import createReactClass from 'create-react-class';
-import Reflux from 'reflux';
-
-import { OverlayTrigger, LinkToNode, Spinner } from 'components/common';
-import { Label } from 'components/bootstrap';
-import InputStateComparator from 'logic/inputs/InputStateComparator';
-import { InputStatesStore } from 'stores/inputs/InputStatesStore';
-import { NodesStore } from 'stores/nodes/NodesStore';
-
-const InputStateBadge = createReactClass({
-  // eslint-disable-next-line react/no-unused-class-component-methods
-  displayName: 'InputStateBadge',
-
-  // eslint-disable-next-line react/no-unused-class-component-methods
-  propTypes: {
-    input: PropTypes.object.isRequired,
-  },
-
-  mixins: [Reflux.connect(InputStatesStore), Reflux.connect(NodesStore)],
-  comparator: new InputStateComparator(),
-
-  _labelClassForState(sortedStates) {
-    const { input } = this.props;
-    const { nodes } = this.state;
-    const nodesWithKnownState = sortedStates.reduce((numberOfNodes, state) => numberOfNodes + state.count, 0);
-
-    if (input.global && nodesWithKnownState !== Object.keys(nodes).length) {
-      return 'warning';
-    }
-
-    const { state } = sortedStates[0];
-
-    switch (state) {
-      case 'RUNNING':
-        return 'success';
-      case 'FAILED':
-      case 'STOPPED':
-        return 'danger';
-      case 'STARTING':
-        return 'info';
-      default:
-        return 'warning';
-    }
-  },
-
-  _textForState(sortedStates) {
-    const { input } = this.props;
-
-    if (input.global) {
-      return sortedStates.map((state) => `${state.count} ${state.state}`).join(', ');
-    }
-
-    return sortedStates[0].state;
-  },
-
-  _isLoading() {
-    const { inputStates, nodes } = this.state;
-
-    return !(inputStates && nodes);
-  },
-
-  render() {
-    if (this._isLoading()) {
-      return <Spinner />;
-    }
-
-    const { input } = this.props;
-    const { inputStates } = this.state;
-    const inputId = input.id;
-    const sortedInputStates = {};
-
-    if (inputStates[inputId]) {
-      Object.keys(inputStates[inputId]).forEach((node) => {
-        const { state } = inputStates[inputId][node];
-
-        if (!sortedInputStates[state]) {
-          sortedInputStates[state] = [];
-        }
-
-        sortedInputStates[state].push(node);
-      });
-    }
-
-    const sorted = Object.keys(sortedInputStates).sort(this.comparator.compare.bind(this.comparator)).map((state) => ({
-      state: state,
-      count: sortedInputStates[state].length,
-    }));
-
-    if (sorted.length > 0) {
-      const popOverText = sorted.map((state) => sortedInputStates[state.state].map((node) => (
-        <small><LinkToNode nodeId={node} />: {state.state}<br />
-        </small>
-      )));
-
-      return (
-        <OverlayTrigger trigger="click" placement="bottom" overlay={popOverText} rootClose title={`Input States for ${input.title}`}>
-          <Label bsStyle={this._labelClassForState(sorted)}
-                 title="Click to show details"
-                 bsSize="xsmall"
-                 style={{ cursor: 'pointer' }}>{this._textForState(sorted)}
-          </Label>
-        </OverlayTrigger>
-      );
-    }
-
-    const text = input.global || input.node === undefined ? '0 RUNNING' : 'NOT RUNNING';
-
-    return (
-      <Label bsStyle="warning" bsSize="xsmall">{text}</Label>
-    );
-  },
-});
-
-export default InputStateBadge;
